backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
sqlparse==0.5.3
Pillow==11.0.0
djangorestframework-simplejwt==5.3.1
django-extensions===3.2.3
Werkzeug===3.1.3
pyOpenSSL===24.3.0
django-admin-inline-paginator==0.4.0
django-mptt==0.17.0
django-image-cropping
easy-thumbnails
django-filter==25.1
openpyxl==3.1.5
unidecode===1.4.0
